+++ main.py
@@ -1,4 +1,6 @@
 import click
+from ClickGroupExtension import SectionedHelpGroup
+import toolbox as hf
 import configparser
 import sympy as sympy
 import wolframalpha
@@ -18,14 +20,15 @@ global_client = wolframalpha.Client(config['WOLFRAMALPHA']['APIKEY'])
 
 # ** This code lacks beauty and is (most probably) inefficient. I had little time. **
 
-@click.group()
+@click.group(cls=SectionedHelpGroup)
 def main():
     """
-    Simple CLI tool for (non) linear programming problems.
+    Simple CLI tool for optimization problems.
     """
     pass
 
-@main.command()
+
+@main.command(help_group='Part 2')
 @click.argument('expression')
 @click.option("--wrt", default='x', help="partial derivative with respect to variable (type 'all' for all)")
 def diffbeauty(expression, wrt):
@@ -40,26 +43,16 @@ def diffbeauty(expression, wrt):
     click.echo("\n\n".join(results))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('expression')
 def difftree(expression):
     """Returns all partial derivatives as a tree."""
-    tree = __difftree_rec(expression)
+    tree = hf.__difftree_rec(expression)
     click.echo(tree.show())
 
 
-@main.command()
-@click.argument('function')
-@click.argument('variable')
-@click.argument('degree')
-def diff(function, variable, degree):
-    """Passes a question to WolframAlpha and returns the answer."""
-    question = "D[(" + function + "), {" + variable + ", " + degree + "}])"
-    res = global_client.query(question)
-    click.echo(next(res.results).text)
 
-
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.argument('substitution')
 def evaluate(function, substitution):
@@ -68,43 +61,43 @@ def evaluate(function, substitution):
     click.echo(next(res.results).text)
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.option("--point", default=None, help="get gradient at point (x,y, ...)")
 @click.option("--value", default=None, help="the value of the point (e.g. '(2,3)'")
 def gradient(function, point, value):
     """Returns the gradient of the given function."""
-    click.echo(__get_gradient(function, point, value))
+    click.echo(hf.__get_gradient(function, point, value))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.option("--det", default=False, help="return determinant of hessian matrix (default is False)")
 def hessian(function, det):
     """Returns Hessian Matrix 'H' of given function."""
-    click.echo(__get_hessian(function, det))
+    click.echo(hf.__get_hessian(function, det))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.argument('substitution')
 def newton(function, substitution):
     """Applies one step of Newton's method."""
     values = []
     for v in substitution.split('=')[1].split(','):
-        values.append(__convert_to_float(str(v).replace('(', '').replace(')', '')))
+        values.append(hf.__convert_to_float(str(v).replace('(', '').replace(')', '')))
     start_vec = np.array([values[0], values[1]])
-    graddres = __get_gradient(function, substitution.split('=')[0], substitution.split('=')[1])
-    gx0 = __convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
-    gy0 = __convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
+    graddres = hf.__get_gradient(function, substitution.split('=')[0], substitution.split('=')[1])
+    gx0 = hf.__convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
+    gy0 = hf.__convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
     grad_vec = np.array([gx0, gy0])
-    H = __get_hessian(function, False)
+    H = hf.__get_hessian(function, False)
     H_res = next(global_client.query("evaluate " + H + "substitute " + substitution).results).text.replace('(',
                                                                                                            '').replace(
         ')', '')
     H_arr_number = []
     for row in [H_res.split('\n')[0].split('|'), H_res.split('\n')[1].split('|')]:
-        desired_array = [__convert_to_float(numeric_string) for numeric_string in row]
+        desired_array = [hf.__convert_to_float(numeric_string) for numeric_string in row]
         H_arr_number.append(desired_array)
     H_inv = np.linalg.inv(H_arr_number)
     new_point = start_vec - H_inv.dot(grad_vec)
@@ -113,7 +106,7 @@ def newton(function, substitution):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.argument('substitution')
 def succhalv(function, substitution):
@@ -123,14 +116,14 @@ def succhalv(function, substitution):
     vars = [x, y]
     values = []
     for v in substitution.split('=')[1].split(','):
-        values.append(__convert_to_float(str(v).replace('(', '').replace(')', '')))
+        values.append(hf.__convert_to_float(str(v).replace('(', '').replace(')', '')))
     refValue = expr.subs(zip(vars, values))
-    graddres = __get_gradient(function, substitution.split('=')[0], substitution.split('=')[1])
+    graddres = hf.__get_gradient(function, substitution.split('=')[0], substitution.split('=')[1])
     x0, y0, B, gx0, gy0 = sympy.symbols('x0, y0, B, gx0, gy0')
     exp1 = x0 + (-B * gx0)
     exp2 = y0 + (-B * gy0)
-    res_exp1 = __convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
-    res_exp2 = __convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
+    res_exp1 = hf.__convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
+    res_exp2 = hf.__convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
     halvings = 0
     f_result = 0
     results = []
@@ -140,16 +133,16 @@ def succhalv(function, substitution):
         y1 = exp2.subs([(y0, values[1]), (B, currentB), (gy0, res_exp2)])
         f_result_previous = f_result
         f_result = expr.subs(zip(vars, [x1, y1]))
-        x1y1 = "({0}, {1})".format(__rstrip_zeros(x1), __rstrip_zeros(y1))
-        results.append([halvings, currentB, x1y1, __rstrip_zeros(f_result), f_result < refValue])
+        x1y1 = "({0}, {1})".format(hf.__rstrip_zeros(x1), hf.__rstrip_zeros(y1))
+        results.append([halvings, currentB, x1y1, hf.__rstrip_zeros(f_result), f_result < refValue])
         if f_result < refValue:
             B_star = currentB / 2 * (3 * refValue - 4 * f_result + f_result_previous) / (
                         refValue - 2 * f_result + f_result_previous)
             x1 = exp1.subs([(x0, values[0]), (B, B_star), (gx0, res_exp1)])
             y1 = exp2.subs([(y0, values[1]), (B, B_star), (gy0, res_exp2)])
-            x1y1 = "({0}, {1})".format(__rstrip_zeros(x1), __rstrip_zeros(y1))
+            x1y1 = "({0}, {1})".format(hf.__rstrip_zeros(x1), hf.__rstrip_zeros(y1))
             f_result_better = expr.subs(zip(vars, [x1, y1]))
-            results.append(['B*', B_star, x1y1, __rstrip_zeros(f_result_better), ' - '])
+            results.append(['B*', B_star, x1y1, hf.__rstrip_zeros(f_result_better), ' - '])
             table = [tabulate(results,
                               headers=["i", "B", "(x1, y1)", "f(x1, y1)", "< " + f"{refValue:.2f}..." + "?"],
                               tablefmt="simple")]
@@ -159,7 +152,7 @@ def succhalv(function, substitution):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('function')
 @click.argument('substitution')
 @click.argument('steps')
@@ -171,7 +164,7 @@ def broyden(function, substitution, steps):
     x_gradient_histroy = []
     values = []
     for v in substitution.split('=')[1].split(','):
-        values.append(__convert_to_float(str(v).replace('(', '').replace(')', '')))
+        values.append(hf.__convert_to_float(str(v).replace('(', '').replace(')', '')))
     start_vec = np.array([values[0], values[1]])
     x_point_histroy.append(start_vec)
     d = []
@@ -180,19 +173,19 @@ def broyden(function, substitution, steps):
     x = values[0]
     y = values[1]
     while step < int(steps):
-        graddres = __get_gradient(function, substitution.split('=')[0], f"({x},{y})")
-        gx0 = __convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
-        gy0 = __convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
+        graddres = hf.__get_gradient(function, substitution.split('=')[0], f"({x},{y})")
+        gx0 = hf.__convert_to_float(graddres.split(',')[0].replace('{', '').replace('}', ''))
+        gy0 = hf.__convert_to_float(graddres.split(',')[1].replace('{', '').replace('}', ''))
         grad_vec = np.array([gx0, gy0])
         x_gradient_histroy.append(grad_vec)
         if step == 0:
             # inverse hessian
-            H = __get_hessian(function, False)
+            H = hf.__get_hessian(function, False)
             H_res = next(global_client.query("evaluate " + H + "substitute " + substitution).results).text \
                 .replace('(', '').replace(')', '')
             H_arr_number = []
             for row in [H_res.split('\n')[0].split('|'), H_res.split('\n')[1].split('|')]:
-                desired_array = [__convert_to_float(numeric_string) for numeric_string in row]
+                desired_array = [hf.__convert_to_float(numeric_string) for numeric_string in row]
                 H_arr_number.append(desired_array)
             A = np.linalg.inv(H_arr_number)
         else:
@@ -215,15 +208,15 @@ def broyden(function, substitution, steps):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('valueseq')
 def aitken(valueseq):
     """Returns the Aitken sequence for a value series of at least 3."""
     vs = valueseq.split(',')
     values = []
     for v in vs:
-        values.append(__convert_to_float(v))
-    if values.__len__() < 3:
+        values.append(hf.__convert_to_float(v))
+    if values.hf.__lenhf.__() < 3:
         click.echo("requires at least 3 values!")
         return
     results = []
@@ -240,17 +233,17 @@ def aitken(valueseq):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.option("--directed", default='False', help="Use 'True' when graph is directed.")
 @click.option("--format", default='html', help="Interactive ('html') or static ('png')")
 def drawgraph(csvfile, directed, format):
     """Plots a graph based on provided adjacency matrix."""
-    G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+    G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
     with open(csvfile, 'r') as f:
         d_reader = csv.DictReader(f)
         headers = d_reader.fieldnames
-    labels = __make_label_dict(headers[1:])
+    labels = hf.__make_label_dict(headers[1:])
     edge_labels = dict(((u, v), d["weight"]) for u, v, d in G.edges(data=True))
     pos = nx.spring_layout(G, k=(5 / (G.order()**(1/2))), iterations=20, scale=5)
     nx.draw(G, pos)
@@ -276,11 +269,11 @@ def drawgraph(csvfile, directed, format):
     click.echo("result saved as: " + file)
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 def mst(csvfile):
     """Returns the minimum spanning tree."""
-    G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+    G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
     T = nx.minimum_spanning_tree(G)
     results = []
     totalweight = 0
@@ -292,12 +285,12 @@ def mst(csvfile):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.argument('fromnode')
 def dijkstra(csvfile, fromnode):
     """All shortest paths to all other nodes from given starting node."""
-    G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+    G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
     p = nx.shortest_path(G, source=fromnode, weight='weight')
     df = pd.DataFrame({'target': p.keys(), 'sp': p.values()})
     results = []
@@ -308,13 +301,13 @@ def dijkstra(csvfile, fromnode):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.argument('fromnode')
 @click.argument('style')
 def traverse(csvfile, fromnode, style):
     """Traverses graph either breadth-first (style='bf') or depth-first (style='df')."""
-    G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+    G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
     if style == 'bf':
         edges = nx.bfs_edges(G, fromnode)
     else:
@@ -333,37 +326,38 @@ def traverse(csvfile, fromnode, style):
     click.echo("\n".join(table) + "\nEncounter Order: " + " → ".join(nodes))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.option("--onlyuse", default='all', help="Node constraints (e.g. 'A, D, F')")
 def floydwarshall(csvfile, onlyuse):
     """Returns matrix with shortest distances between all nodes."""
-    G, m, labels = __get_graph_from_adjacency_matrix(csvfile, filling_values=np.inf)
+    G, m, labels = hf.__get_graph_from_adjacency_matrix(csvfile, filling_values=np.inf)
     if onlyuse == 'all':
         #p = nx.floyd_warshall_numpy(G, weight='weight')
         allowed_indexes = range(np.size(m[0]))
     else:
         allowed_indexes = [i for i, x in enumerate(labels) if x in np.char.strip(onlyuse.split(','))]
 
-    p, change = __floydwarshall_constrained(m, allowed_indexes)
+    p, change = hf.__floydwarshall_constrained(m, allowed_indexes)
     results = np.c_[list(G.nodes), p, np.repeat(' | ', np.size(m[0])), change]
     headers = np.hstack((list(G.nodes), [' | '], list(G.nodes)))
     table = [tabulate(results, headers=headers, tablefmt="simple", stralign="center")]
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.argument('source')
 @click.argument('target')
 def maxflow(csvfile, source, target):
     """Finds maximum flow based on provided edge list."""
-    G = __get_graph_from_edge_list(csvfile)
+    G = hf.__get_graph_from_edge_list(csvfile)
     maxflow, transactions = nx.maximum_flow(G, source, target, 'weight')
     table = [tabulate(np.array(list(transactions.items())), headers=["node", "routed values"], tablefmt="simple")]
     click.echo("max flow: " + str(maxflow) + "\n" + "\n".join(table))
 
-@main.command()
+
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.argument('source')
 @click.argument('target')
@@ -371,10 +365,10 @@ def maxflow(csvfile, source, target):
 def mincut(csvfile, source, target, adjacency):
     """Finds minimum s-t-cut based on provided edge list or adjacency matrix."""
     if adjacency:
-        G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+        G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
         cut_value, partition = nx.minimum_cut(G, source, target, capacity='weight')
     else:
-        G = __get_graph_from_edge_list(csvfile)
+        G = hf.__get_graph_from_edge_list(csvfile)
         cut_value, partition = nx.minimum_cut(G, source, target, capacity='weight')
     results = []
     i = 0
@@ -386,11 +380,11 @@ def mincut(csvfile, source, target, adjacency):
 
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 def maxmatch(csvfile):
     """Maximum matchings of a bipartite graph based on provided adjacency matrix."""
-    G, _, _ = __get_graph_from_adjacency_matrix(csvfile)
+    G, _, _ = hf.__get_graph_from_adjacency_matrix(csvfile)
     matching = nx.maximal_matching(G)
     results = []
     for m in list(matching):
@@ -399,13 +393,13 @@ def maxmatch(csvfile):
     click.echo("\n".join(table))
 
 
-@main.command()
+@main.command(help_group='Part 2')
 @click.argument('csvfile')
 @click.argument('source')
 @click.argument('target')
 def mincostmaxflow(csvfile, source, target):
     """Returns a maximum s-t flow of minimum cost based on provided edge list."""
-    G = __get_graph_from_edge_list(csvfile)
+    G = hf.__get_graph_from_edge_list(csvfile)
     flowDict = nx.max_flow_min_cost(G, source, target, capacity='weight', weight='cost')
     mincost = nx.cost_of_flow(G, flowDict, weight='cost')
     mincostFlowValue = sum((flowDict[u][target] for u in G.predecessors(target))) - sum((flowDict[target][v] for v in G.successors(target)))
@@ -414,109 +408,6 @@ def mincostmaxflow(csvfile, source, target):
                "max flow: " + str(mincostFlowValue) + "\n" +
                "\n".join(table))
 
-
-def __convert_to_float(frac_str):
-    try:
-        if frac_str.__contains__('^'):
-            frac_str = frac_str.replace('×10^', 'E')
-        return float(frac_str)
-    except ValueError:
-        num, denom = frac_str.split('/')
-        try:
-            leading, num = num.split(' ')
-            whole = float(leading)
-        except ValueError:
-            whole = 0
-        frac = float(num) / float(denom)
-        return whole - frac if whole < 0 else whole + frac
-
-def __rstrip_zeros(f):
-    return ('%f' % f).rstrip('0').rstrip('.')
-
-def __difftree_rec(expression, level=0, functionsuffix='', transform=True, parentNode='root', tree=None):
-    if transform:
-        expr = sympy.parse_expr(expression, transformations=sympy.parsing.sympy_parser.T[:])
-    else:
-        expr = expression
-    if level == 0:
-        tree = Tree()
-        tree.create_node("f: {0}".format(str(expr).replace('**', '^').replace('*', '')), 'root')
-        level += 1
-    for v in expr.free_symbols:
-        d = sympy.diff(expr, v)
-        id = "f{0}{1}_{2}".format(level, v, functionsuffix)
-        tree.create_node("{0}: {1}".format(id, str(d).replace('**', '^').replace('*', '')), id, parent=parentNode)
-        if not d.is_constant():
-            __difftree_rec(d.as_expr(), level + 1,
-                           functionsuffix=(functionsuffix + str(v)),
-                           transform=False,
-                           parentNode=id, tree=tree)
-    return tree
-
-def __get_gradient(function, point, value):
-    question = "grad (" + function + ")"
-    res = global_client.query(question)
-    gradient = res.details['Result in 2D Cartesian coordinates'].split("\n")[0]
-    returnValue = gradient
-    if point is not None:
-        res2 = global_client.query("evaluate " + gradient.split("=")[1] + "substitute " + point + " = " + value)
-        returnValue = next(res2.results).text
-    return returnValue
-
-def __get_hessian(function, det):
-    if det:
-        question = "hessian of (" + function + ")"
-        res = global_client.query(question)
-        result = next(res.results).text
-    else:
-        question = "hessian matrix of (" + function + ")"
-        res = global_client.query(input=question, format='moutput')
-        result = ((next(res.results))['subpod'])['moutput']
-    return result
-
-def __make_label_dict(labels):
-    l = {}
-    for i, label in enumerate(labels):
-        l[i] = label
-    return l
-
-def __get_graph_from_adjacency_matrix(csvfile, filling_values=None):
-    with open(csvfile, 'r') as f:
-        ncols = len(next(f).split(','))
-    x = np.genfromtxt(csvfile, delimiter=',', filling_values=filling_values, dtype='float32', names=True, usecols=range(1, ncols))
-    labels = x.dtype.names
-    y = x.view(dtype=('float32', len(x.dtype)))
-    G = nx.from_numpy_matrix(y)
-    G = nx.relabel_nodes(G, dict(zip(range(ncols - 1), labels)))
-    return G, y, list(labels)
-
-def __get_graph_from_edge_list(csvfile, directed = True):
-    xs = np.genfromtxt(csvfile, delimiter=',', dtype=None, names=True)
-    edges = []
-    for x in xs:
-        fnode = x[0]
-        if isinstance(fnode, bytes):
-            fnode = fnode.decode("utf-8")
-        tnode = x[1]
-        if isinstance(tnode, bytes):
-            tnode = tnode.decode("utf-8")
-        edges.append((fnode, tnode, {"weight": x[2], "cost": x[3]}))
-    if directed:
-        G = nx.DiGraph()
-    else:
-        G = nx.Graph()
-    G.add_edges_from(edges)
-    return G
-
-def __floydwarshall_constrained(m, allowed_indexes):
-    n = len(m[1])
-    m_old = np.copy(m)
-    m_old[np.isinf(m_old)] = 0
-    for k in allowed_indexes:
-        for i in range(n):
-            for j in range(n):
-                m[i, j] = min(m[i, j], m[i, k]+m[k, j])
-    return m, m - m_old
 
 
 if __name__ == "__main__":
